Refactor: apply DndInterface refactor to Tutorial
Refactor: separate sample data as sampleData.js
